Moving exchangeToken to its proper place
