+++ src/app/(landing)/oauth-callback/facebook/page.tsx
@@ -9,8 +9,8 @@ import { Loader } from "lucide-react"
 import { Card, CardHeader, CardTitle } from "@/components/ui/card"
 import { Icons } from "@/components/icons"
 
+import { exchangeToken } from "../../../../actions/turnkey"
 import { verifierSegmentToChallenge } from "../../../../lib/facebook-utils"
-import { exchangeToken } from "./exchange-token"
 
 function FacebookProcessCallback() {
   const searchParams = useSearchParams()
